feat: 주별 식단 캘린더 프레임 작성
