+++ package.json
@@ -16,6 +16,8 @@
     "font-awesome": "^4.7.0",
     "moment": "^2.29.4",
     "openvidu-browser": "2.22.0",
+    "rc-menu": "^9.6.4",
+    "rc-select": "^14.1.13",
     "react": "^18.2.0",
     "react-apexcharts": "^1.4.0",
     "react-big-calendar": "^1.5.0",
